+++ lib/opal/nodes/top.rb
@@ -15,8 +15,6 @@ module Opal
 
         opening
         in_scope do
-          compile_config
-
           body_code = stmt(stmts)
           body_code = [body_code] unless body_code.is_a?(Array)
 
@@ -98,15 +96,6 @@ module Opal
           line "var $__END__ = Opal.Object.$new();"
           line "$__END__.$read = function() { return #{content.inspect}; };"
         end
-      end
-
-      def compile_config
-        line "var OPAL_CONFIG = { "
-        push "method_missing: #{compiler.method_missing?}, "
-        push "arity_check: #{compiler.arity_check?}, "
-        push "freezing: #{compiler.freezing?}, "
-        push "tainting: #{compiler.tainting?} "
-        push "};"
       end
 
       def version_comment